adicionando projeteis(finalizado)

+++ Projetil/Projetil.h
@@ -4,19 +4,20 @@
 #include <vector>
 #include <utility>
 #include "../Protagonista/Protagonista.h"
+#include "../Parede/Parede.h"
 
 using namespace std;
 
 class Projetil {
     private:
-        pair<int,int> posicao;
+        pair<float,float> posicao;
         pair<int,int> tamanho;
-        pair<int,int> move_speed;
+        pair<float,float> move_speed;
         int dano;
         bool ehInimigo;
     public:
         int ricochetes;
-        Projetil(pair<int,int> p, pair<int,int> t, pair<int,int> ms, int d, int r, bool ei);
+        Projetil(pair<float,float> p, pair<int,int> t, pair<float,float> ms, int d, int r, bool ei);
 
         int getX();
         int getY();
@@ -24,7 +25,7 @@ class Projetil {
         int getTamanhoX();
         int getTamanhoY();
         
-        pair<int,int> getMS();
+        pair<float,float> getMS();
         int getDano();
         void ricochetear(Parede par);
